Latest updates
Updated all the npm package dependencies
- added eslint-disable-next-line no-console for any console logs I do want to print out still - these are for error logging in the console. This is because I have a rule setup on my linter that warns of any console logs.
- NEXT UI MODAL CHANGES
Since updating all the npm packages, Next UI's modal ("^2.0.36") size was not being respected when in sm: screen breakpoint.
It's to do with a height and min height of 100dvh (100% of the device's height) being set by default, when selecting the prop: 'size="full".
 I now have to override the h and min-h when moving to sm: breakpoint. Edits to the two modals I have in the app have been made.